# docs/openapi.yaml
openapi: 3.0.3
...
info:
  title: Mova – Gym Tracker API
  description: |
    REST API for the **Mova** gym-tracking web application.

    ## Authentication
    All endpoints (except registration) require **HTTP Basic Auth**.
    Include your credentials via the `Authorization: Basic <base64(username:password)>` header.

    ## Roles
    | Role    | Description |
    |---------|-------------|
    | `admin` | Manages exercises, templates, and exercise requests |
    | `user`  | Logs workouts, tracks metrics, manages personal plans |

    ## Business Rules
    1. **Exercise Authority Integrity** — Only admins can create/update/delete exercises. Deletion is a soft-delete (`is_active = false`) to preserve historical data.
    2. **Anti-Spam Security** — A user can have at most **5 pending** exercise requests at any time.
    3. **Data Validation** — A workout session cannot be saved unless it contains at least one workout set.
  version: 1.0.0
  contact:
    name: Team Mova
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.mova.app
    description: Production

# =============================================================================
#  Global Security — Basic Auth on every endpoint by default
# =============================================================================
security:
  - basicAuth: []
...
paths:
...
  /exercises:
    get:
      tags: [Exercises]
      summary: List all exercises
      description: |
        Returns all exercises from the master catalogue.
        By default only active exercises are returned. Admins can include inactive exercises using the `include_inactive` filter.
      operationId: listExercises
      parameters:
        - name: muscle_group
          in: query
          required: false
          description: Filter by muscle group
          schema:
            $ref: "#/components/schemas/MuscleGroup"
        - name: is_cardio
          in: query
          required: false
          description: Filter by cardio flag
          schema:
            type: boolean
        - name: include_inactive
          in: query
          required: false
          description: "**Admin only.** If `true`, also returns soft-deleted exercises."
          schema:
            type: boolean
            default: false
        - name: search
          in: query
          required: false
          description: Search exercises by name (case-insensitive partial match)
          schema:
            type: string
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Exercises]
      summary: Create a new exercise
      description: "**Admin only.** Adds a new exercise to the master catalogue."
      operationId: createExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseCreate"
      responses:
        "201":
          description: Exercise created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: An exercise with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /exercises/{exerciseId}:
    get:
      tags: [Exercises]
      summary: Get an exercise by ID
      operationId: getExercise
      parameters:
        - $ref: "#/components/parameters/ExerciseIdParam"
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Exercises]
      summary: Update an exercise
      description: "**Admin only.** Updates exercise details (name, muscle group, cardio flag)."
      operationId: updateExercise
      parameters:
        - $ref: "#/components/parameters/ExerciseIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseCreate"
      responses:
        "200":
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: An exercise with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [Exercises]
      summary: Soft-delete an exercise
      description: |
        **Admin only.** Sets `is_active = false` on the exercise.
        The exercise remains in the database so historical workout data is preserved (Business Rule 1).
      operationId: deleteExercise
      parameters:
        - $ref: "#/components/parameters/ExerciseIdParam"
      responses:
        "200":
          description: Exercise soft-deleted (returned with `is_active = false`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  #  EXERCISE REQUESTS  (User-submitted suggestions)
  # ---------------------------------------------------------------------------
  /exercise-requests:
    get:
      tags: [Exercise Requests]
      summary: List exercise requests
      description: |
        - **Admin:** Returns all exercise requests (optionally filtered by status).
        - **User:** Returns only their own exercise requests.
      operationId: listExerciseRequests
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by request status
          schema:
            $ref: "#/components/schemas/RequestStatus"
      responses:
        "200":
          description: List of exercise requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Exercise Requests]
      summary: Submit a new exercise request
      description: |
        **User only.** Submits a request for a new exercise to be added.
        The request starts in `pending` status.
        A user cannot have more than **5 pending** requests at any time (Business Rule 2).
      operationId: createExerciseRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseRequestCreate"
      responses:
        "201":
          description: Exercise request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          description: "User has reached the maximum of 5 pending requests (Business Rule 2)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /exercise-requests/{requestId}:
    get:
      tags: [Exercise Requests]
      summary: Get an exercise request by ID
      description: "**Admin or the requesting user.**"
      operationId: getExerciseRequest
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Exercise request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Exercise Requests]
      summary: Delete an exercise request
      description: "**Admin or the requesting user.** Only `pending` requests can be deleted."
      operationId: deleteExerciseRequest
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "204":
          description: Exercise request deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot delete a request that has already been approved or denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /exercise-requests/{requestId}/approve:
    post:
      tags: [Exercise Requests]
      summary: Approve an exercise request
      description: |
        **Admin only.** Atomically approves the request and creates the
        corresponding exercise in the master catalogue.

        The request status transitions from `pending` to `approved`, and
        a new `Exercise` is created using the request's `suggested_name`,
        `muscle_group`, and `is_cardio` values.

        If an exercise with the same name already exists, the approval
        fails with `409 Conflict` and no changes are persisted.
      operationId: approveExerciseRequest
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Request approved and exercise created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Either (a) the request is not in `pending` status, or
            (b) an exercise with the same name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /exercise-requests/{requestId}/deny:
    post:
      tags: [Exercise Requests]
      summary: Deny an exercise request
      description: "**Admin only.** Sets the request status to `denied`."
      operationId: denyExerciseRequest
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Request denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Request is not in `pending` status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  #  WORKOUT PLANS  (Templates & personal routines)
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: "HTTP Basic Authentication. Send `Authorization: Basic <base64(username:password)>`."

  # ---------------------------------------------------------------------------
  #  Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    ExerciseIdParam:
      name: exerciseId
      in: path
      required: true
      description: Unique identifier of the exercise
      schema:
        type: integer
    RequestIdParam:
      name: requestId
      in: path
      required: true
      description: Unique identifier of the exercise request
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but insufficient permissions for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ---------------------------------------------------------------------------
  #  Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    MuscleGroup:
      type: string
      description: Target muscle group for an exercise
      enum:
        - Chest
        - Back
        - Legs
        - Shoulders
        - Arms
        - Core
        - Cardio
    RequestStatus:
      type: string
      description: Lifecycle status of an exercise request
      enum:
        - pending
        - approved
        - denied
...
    Exercise:
      type: object
      description: A master exercise definition (Aggregate Root)
      properties:
        id:
          type: integer
          readOnly: true
          example: 42
        name:
          type: string
          example: Barbell Squat
        muscle_group:
          $ref: "#/components/schemas/MuscleGroup"
        is_cardio:
          type: boolean
          example: false
        is_active:
          type: boolean
          description: "`false` means the exercise has been soft-deleted"
          example: true
          readOnly: true
      required:
        - id
        - name
        - muscle_group
        - is_cardio
        - is_active
    ExerciseCreate:
      type: object
      description: Payload for creating or updating an exercise
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Barbell Squat
        muscle_group:
          $ref: "#/components/schemas/MuscleGroup"
        is_cardio:
          type: boolean
          example: false
      required:
        - name
        - muscle_group
        - is_cardio

    # ── Exercise Request Aggregate ─────────────────────────────────────────
    ExerciseRequest:
      type: object
      description: A user-submitted suggestion for a new exercise (Aggregate Root)
      properties:
        id:
          type: integer
          readOnly: true
          example: 7
        user_id:
          type: integer
          readOnly: true
          example: 1
        suggested_name:
          type: string
          example: Bulgarian Split Squat
        muscle_group:
          $ref: "#/components/schemas/MuscleGroup"
        is_cardio:
          type: boolean
          example: false
        status:
          $ref: "#/components/schemas/RequestStatus"
      required:
        - id
        - user_id
        - suggested_name
        - muscle_group
        - is_cardio
        - status
    ExerciseRequestCreate:
      type: object
      description: Payload for submitting a new exercise request
      properties:
        suggested_name:
          type: string
          minLength: 2
          maxLength: 100
          example: Bulgarian Split Squat
        muscle_group:
          $ref: "#/components/schemas/MuscleGroup"
        is_cardio:
          type: boolean
          description: Whether this is a cardio exercise
          example: false
      required:
        - suggested_name
        - muscle_group
        - is_cardio

    # ── Workout Plan Aggregate ─────────────────────────────────────────────
...
    Error:
      type: object
      description: Standard error response
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          description: Human-readable error description
          example: The requested resource was not found
      required:
        - error
        - message
    ValidationErrorResponse:
      type: object
      description: Detailed validation error response
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
          example: Request body failed validation
        details:
          type: array
          description: List of individual field validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: "Field is required"
      required:
        - error
        - message
        - details